TIMER1: Correcting definitions of hardware registers from pointer to u32 to u8

+++ lib/TIMER1_private.h
@@ -2,14 +2,14 @@
 #define _TIMER1_PRIVATE_H
 
 /*Register definitions*/
-#define TCCR1A *((u32*) 0x4F)
-#define TCCR1B *((u32*) 0x4E)
-#define OCR1AH *((u32*) 0x4A)
-#define OCR1AL *((u32*) 0x4B)
-#define OCR1A * ((u32*) 0x4B)
-#define OCR1BH *((u32*) 0x49)
-#define OCR1BL  *((u32*) 0x48)
-#define OCR1B * ((u32*) 0x48)
+#define TCCR1A *((u8*) 0x4F)
+#define TCCR1B *((u8*) 0x4E)
+#define OCR1AH *((u8*) 0x4A)
+#define OCR1AL *((u8*) 0x4B)
+#define OCR1A * ((u8*) 0x4B)
+#define OCR1BH *((u8*) 0x49)
+#define OCR1BL  *((u8*) 0x48)
+#define OCR1B *((u8*) 0x48)
 
 
 #endif
